Update LXC default.conf to ignore incomplete AppArmor configuration
When LXC runs in a docker container on a host that has AppArmor (AA)
enabled, e.g. Ubuntu, it causes LXC to error out because it detects the
presence of AA via the kernel, but it does not see AA configured
properly in the docker container. To ignore this error, tell LXC to
allow incomplete AA configurations.

Change-Id: Ib457531d9a67530ead6e33c88e528d66f09ac83b